+++ example/linux/flutter/generated_plugins.cmake
@@ -4,6 +4,7 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   file_picker_pro
+  file_selector_linux
   open_share_plus
   url_launcher_linux
 )