+++ lib/files.dart
@@ -26,6 +26,42 @@ enum FileMode {
 class Files {
   Files._();
 
+  static const String txt = "txt";
+  static const String png = "png";
+  static const String jpg = "jpg";
+  static const String jpeg = "jpeg";
+  static const String bmp = "bmp";
+  static const String svg = "svg";
+  static const String gif = "gif";
+  static const String pdf = "pdf";
+  static const String doc = "doc";
+  static const String docx = "docx";
+  static const String docm = "docm";
+  static const String xls = "xls";
+  static const String xlsx = "xlsx";
+  static const String xlsm = "xlsm";
+  static const String csv = "csv";
+  static const String ppt = "ppt";
+  static const String pptx = "pptx";
+  static const String pptm = "pptm";
+  static const String zip = "zip";
+  static const String rar = "rar";
+  static const String jar = "jar";
+  static const String odt = "odt";
+  static const String ods = "ods";
+  static const String apk = "apk";
+  static const String ipa = "ipa";
+  static const String mp3 = "mp3";
+  static const String mp4 = "mp4";
+  static const String mpeg = "mpeg";
+  static const String mpg = "mpg";
+  static const String m4a = "m4a";
+  static const String mov = "mov";
+  static const String wav = "wav";
+  static const String rtf = "rtf";
+  static const String otf = "otf";
+  static const String ttf = "ttf";
+
   static const String cropperToolbarTitle = "Crop";
   static const Color cropperToolbarColor = Colors.black;
   static const Color cropperToolbarWidgetsColor = Colors.white;
@@ -37,33 +73,6 @@ class Files {
       "File more than $maxFileSizeInMb MB, Please select another file";
   static const int _mcFPCForCancel = 1;
   static const int _mcFPCForSize = 2;
-
-  /// file extensions
-  static const pdf = "pdf";
-  static const doc = "doc";
-  static const docx = "docx";
-  static const xls = "xls";
-  static const xlsx = "xlsx";
-  static const ppt = "ppt";
-  static const pptx = "pptx";
-  static const jpg = "jpg";
-  static const jpeg = "jpeg";
-  static const png = "png";
-  static const gif = "gif";
-
-  /// allowed all extensions
-  static const List<String> allowedAllExtensions = [
-    Files.pdf,
-    Files.doc,
-    Files.docx,
-    Files.xls,
-    Files.xlsx,
-    Files.ppt,
-    Files.pptx,
-    Files.png,
-    Files.jpg,
-    Files.jpeg
-  ];
 
   static bool _isNullOREmpty(String? str) {
     if (str == null || str.isEmpty) {
@@ -217,7 +226,7 @@ class Files {
       String cropperToolbarTitle = Files.cropperToolbarTitle,
       Color cropperToolbarColor = Files.cropperToolbarColor,
       Color cropperToolbarWidgetsColor = Files.cropperToolbarWidgetsColor,
-      List<String> allowedExtensions = Files.allowedAllExtensions}) async {
+      List<String>? allowedExtensions}) async {
     fileMode == FileMode.camera
         ? await Files.cameraPicker(
             fileData: fileData,
@@ -398,18 +407,7 @@ class Files {
       required Function(FileData fileData) onSelected,
       Function(String message, int messageCode)? onCancel,
       int? maxFileSizeInMb,
-      List<String> allowedExtensions = const [
-        Files.pdf,
-        Files.doc,
-        Files.docx,
-        Files.xls,
-        Files.xlsx,
-        Files.ppt,
-        Files.pptx,
-        Files.png,
-        Files.jpg,
-        Files.jpeg
-      ]}) async {
+      List<String>? allowedExtensions}) async {
     FilePickerResult? result = await FilePicker.platform
         .pickFiles(type: FileType.custom, allowedExtensions: allowedExtensions);
     if (result != null && result.files.single.path != null) {