+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.4.0
 
-* Flutter plugin for selecting files from any devices file library, and taking a new pictures with the camera.
+* Include more extensions.
 
 ## 1.3.0